Fix test: test_converter.TestConverter.test_volume_is_indexed could only run alone.
If it ran after `test_does_not_index_volume_properties_for_articles` it broke
because that test changed global data structures in VolumeIndex (aka the `solr` attribute), so the volume index got
omitted in `SolrConverter.convert()` because of the duplicate `product_id` name.

This seems to be broken live, too as indexing a Volume without `date_digital_published` no longer renders this field for all following index calls as the the index has a name which already exists in `SolrConverter.solr_fields_seen`.